# contracts/openapi/public.v1.yaml
openapi: 3.1.0
info:
  title: RagSpace Public API
  version: 1.0.0
  summary: The only API the browser talks to.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    Served by `api-gateway` under `/api/v1`. The gateway authenticates the
    caller, enforces rate limits and translates each operation into gRPC calls
    to internal services (see `contracts/proto`). Internal services are never
    reachable from the browser.

    ## Conventions
    - **JSON.** Field names are `camelCase`. Timestamps are ISO-8601 UTC.
      Sizes are `…Bytes`, durations `…Seconds`. IDs are opaque strings.
    - **Nulls.** Responses always include every documented field, using
      `null` for absent values. Requests may omit optional fields.
    - **Strict input, tolerant output.** Request bodies with unknown fields
      are rejected (`400`). Clients must ignore unknown response fields and
      handle unknown enum values, so additive changes are not breaking.
    - **Versioning.** Additive changes ship within `v1`. Breaking changes ship
      as `/api/v2`.
    - **Enums.** Domain enum values are `UPPER_SNAKE_CASE`. Discriminators
      (`type`, `kind`) and event names are lower case.
    - **PATCH** bodies follow JSON Merge Patch semantics (RFC 7396): an omitted
      field is unchanged, `null` clears it.
    - **Creation.** `201 Created` with a `Location` header. Non-idempotent
      POSTs accept an `Idempotency-Key` header; retries with the same key
...
servers:
...
security:
  - sessionCookie: []
...
paths:
...
  /auth/sign-up:
    post:
      tags: [Auth]
      operationId: signUp
      summary: Create an account with email and password
      description: Signs the user in on success. Rate limit 5/min per IP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '201':
          description: Account created and session started.
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SetSessionCookie'
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/ServerError'
  /auth/sign-in:
    post:
      tags: [Auth]
      operationId: signIn
      summary: Sign in with email and password
      description: Rate limit 10/min per IP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInRequest'
      responses:
        '200':
          description: Session started.
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SetSessionCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/ServerError'
  /auth/sign-out:
    post:
      tags: [Auth]
      operationId: signOut
      summary: End the current session
      description: Idempotent; succeeds even without a session.
      security:
        - {}
        - sessionCookie: []
      responses:
        '204':
          description: Session revoked and cookie cleared.
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SetSessionCookie'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/ServerError'
  /auth/oauth/{provider}:
    get:
      tags: [Auth]
      operationId: startOAuth
      summary: Start an OAuth sign-in
      security: []
      parameters:
        - $ref: '#/components/parameters/OAuthProvider'
        - name: redirectTo
          in: query
          description: Same-origin path to land on after sign-in.
          schema:
            type: string
            pattern: '^/(?![/\\])'
            maxLength: 512
            default: /
      responses:
        '302':
          description: Redirect to the provider's consent screen.
          headers:
            Location:
              $ref: '#/components/headers/Location'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/ServerError'
...
  /auth/password/forgot:
    post:
      tags: [Auth]
      operationId: requestPasswordReset
      summary: Email a password reset link
      description: |
        Always `202`, whether or not the email exists, so it cannot be used to
        discover accounts. Rate limit 3/min per IP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '202':
          description: Accepted.
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/ServerError'
  /auth/password/reset:
    post:
      tags: [Auth]
      operationId: resetPassword
      summary: Set a new password with a reset token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          description: Password changed; all sessions revoked.
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/ServerError'
  /auth/password/change:
    post:
      tags: [Auth]
      operationId: changePassword
      summary: Change password while signed in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/ServerError'
  /me:
    get:
      tags: [Account]
      operationId: getMe
      summary: The signed-in user
      description: Also the way to check whether a session is active.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/ServerError'
    patch:
      tags: [Account]
      operationId: updateMe
      summary: Update profile fields
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeRequest'
      responses:
        '200':
          description: Updated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/ServerError'
    delete:
      tags: [Account]
      operationId: deleteMe
      summary: Delete the account and all its data
      description: |
        Requires a fresh session (signed in within the last 10 minutes);
        otherwise `403`. The account is removed immediately and the cookie
        cleared. Files, embeddings, scenes, conversations and the subscription
        are purged asynchronously.
      responses:
        '202':
          description: Accepted.
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SetSessionCookie'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/ServerError'
...
  /webhooks/{provider}:
    post:
      tags: [Webhooks]
      operationId: receivePaymentWebhook
      summary: Payment provider callback
      description: |
        Server to server; exempt from the `Origin` check. The raw body and
        headers are forwarded to payment-service, which verifies the provider
        signature (`X-Signature` for Lemon Squeezy, `X-Razorpay-Signature` for
        Razorpay). Redeliveries of the same provider event are no-ops.
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [lemon-squeezy, razorpay]
        - name: X-Signature
          in: header
          schema:
            type: string
            maxLength: 512
        - name: X-Razorpay-Signature
          in: header
          schema:
            type: string
            maxLength: 512
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '204':
          description: Processed.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/ServerError'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: better-auth.session_token
  headers:
    SetSessionCookie:
      description: Sets or clears the httpOnly session cookie.
      schema:
        type: string
    Location:
      description: URL of the created or target resource.
      schema:
        type: string
        format: uri-reference
    RetryAfter:
      description: Seconds until the client may retry.
      schema:
        type: integer
        format: int32
        minimum: 0
    RateLimitLimit:
      description: Requests allowed in the current window.
      schema:
        type: integer
        format: int32
        minimum: 0
    RateLimitRemaining:
      description: Requests left in the current window.
      schema:
        type: integer
        format: int32
        minimum: 0
    RateLimitReset:
      description: Seconds until the window resets.
      schema:
        type: integer
        format: int32
        minimum: 0
  parameters:
...
    OAuthProvider:
      name: provider
      in: path
      required: true
      schema:
        type: string
        enum: [google]
  responses:
    BadRequest:
      description: Invalid request (`validation_failed`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://docs.ragspace.app/errors/validation_failed
            title: Validation failed
            status: 400
            detail: One or more fields are invalid.
            code: validation_failed
            correlationId: 3f1c2a9e-8b7d-4c1e-9f0a-2b6d5e4c3a21
            errors:
              - field: name
                message: must not be empty
    Unauthorized:
      description: Missing or invalid session, or bad credentials (`unauthenticated`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: |
        Request refused (`forbidden`): failed `Origin` check, or the operation
        needs a fresher session.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Conflict:
      description: Current state does not allow the request (`conflict`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    TooManyRequests:
      description: Rate limited (`rate_limited`).
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
        RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ServerError:
      description: |
        Unexpected failure (`internal`), a dependency is down
        (`upstream_unavailable`, 503) or timed out (`upstream_timeout`, 504).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Id:
      type: string
      minLength: 1
      maxLength: 64
      pattern: '^[A-Za-z0-9_-]+$'
    Problem:
      type: object
      description: |
        RFC 9457 problem details plus a stable machine-readable `code`.
        Extension members (`errors`, `metric`, `used`, `limit`) appear only
        where noted.
      required: [type, title, status, code, correlationId]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
        detail:
          type: string
        code:
          type: string
          enum:
            - validation_failed
            - unauthenticated
            - forbidden
            - not_found
            - conflict
            - quota_exceeded
            - rate_limited
            - upstream_unavailable
            - upstream_timeout
            - internal
        correlationId:
          type: string
        errors:
          description: Field-level details for `validation_failed`.
          type: array
          maxItems: 100
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
        metric:
          $ref: '#/components/schemas/UsageMetric'
        used:
          type: integer
          format: int64
          minimum: 0
        limit:
          type: integer
          format: int64
          minimum: 0
...
    Password:
      type: string
      minLength: 8
      maxLength: 128
    SignUpRequest:
      type: object
      additionalProperties: false
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 254
        password:
          $ref: '#/components/schemas/Password'
    SignInRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          maxLength: 128
    ForgotPasswordRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
    ResetPasswordRequest:
      type: object
      additionalProperties: false
      required: [token, newPassword]
      properties:
        token:
          type: string
          maxLength: 512
        newPassword:
          $ref: '#/components/schemas/Password'
    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          maxLength: 128
        newPassword:
          $ref: '#/components/schemas/Password'
        revokeOtherSessions:
          type: boolean
          default: true
    AuthResult:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      required: [id, email, name, username, imageUrl, emailVerified, createdAt, updatedAt]
      properties:
        id:
          $ref: '#/components/schemas/Id'
        email:
          type: string
          format: email
        name:
          type: string
        username:
          type: [string, 'null']
        imageUrl:
          type: [string, 'null']
          format: uri
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateMeRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        username:
          type: [string, 'null']
          maxLength: 30
          pattern: '^[a-zA-Z0-9_]{3,30}$'
...
    UsageMetric:
      type: string
      enum:
        - CONVERSATIONS
        - STORAGE
        - FILE_CONVERSATIONS
        - YOUTUBE_VIDEOS
        - MAX_VIDEO_LENGTH
        - MAX_AUDIO_DURATION
...